DROTH-2098 Unknown speed limit list upgrade part 2 -Changed class name to a more suggestive name -Changed Dyn Sessiont to Transaction -Validate unknownSpeedLimitIds -Changed method of validating if in right worklist to put the delete button -Fixed a bug where it appears two or more buttons in the worklist

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/dao/linearasset/OracleSpeedLimitDao.scala
@@ -526,7 +526,7 @@ class OracleSpeedLimitDao(val vvhClient: VVHClient, val roadLinkService: RoadLin
     sqlu"""delete from unknown_speed_limit where link_id in (#${linkIds.mkString(",")})""".execute
   }
 
-  def hideUnknownSpeedLimits(linkIds: Set[Int]): Set[Int] = {
+  def hideUnknownSpeedLimits(linkIds: Set[Long]): Set[Long] = {
     sqlu"""update unknown_speed_limit set unnecessary = 1 where link_id in (#${linkIds.mkString(",")})""".execute
     linkIds
   }
